trustguard-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
tldextract
pymongo
python-multipart
firecrawl-py
beautifulsoup4
validators
